Added the fillDirection
The fillDirection feature works in conjunction with adjustViewBounds to ensure that the image is fitted along the specified direction, and that the other direction of the view is resized to ensure the whole image is visible. One can also use the fillDimension XML attribute to set it.

Added a test case for the fillDirection=HORIZONTAL.

Changed the way the empty drawable is assigned: it's no longer through the setBackground*() methods, but through its own dedicated setEmptyDrawable method. One can also use the emptyDrawable XML attribute.

Created an IDEA project so that one can now open the ImageViewEx project both in standalone mode and in module mode (when used as library)

Small fixes here and there